* elf-m10200.c (mn10200_elf_relax_section): Cast assignment to 	Elf_Internal_Shdr.contents now that it's no longer a PTR. 	* elf-m10300.c (mn10300_elf_relax_section): Likewise. 	* elf32-h8300.c (elf32_h8_relax_section): Likewise. 	* elf32-m68k.c (bfd_m68k_elf32_create_embedded_relocs): Likewise. 	* elf32-sh.c (sh_elf_relax_section): Likewise. 	* elf64-alpha.c (elf64_alpha_relax_section): Likewise. 	* elfxx-ia64.c (elfNN_ia64_relax_section): Likewise. 	* elf.c (setup_group): Warning fixes. 	* elflink.h (elf_link_sort_relocs): Likewise. 	* pdp11.c (slurp_reloc_table): Likewise.

+++ bfd/elflink.h
@@ -4480,7 +4480,7 @@ elf_link_sort_relocs (abfd, info, psec)
 	  }
       }
 
-  qsort (rela, count, sizeof (*rela), elf_link_sort_cmp1);
+  qsort (rela, (size_t) count, sizeof (*rela), elf_link_sort_cmp1);
   for (ret = 0; ret < count && rela[ret].type == reloc_class_relative; ret++)
     ;
   for (i = ret, j = ret; i < count; i++)
@@ -4489,7 +4489,7 @@ elf_link_sort_relocs (abfd, info, psec)
 	j = i;
       rela[i].offset = rela[j].u.rel.r_offset;
     }
-  qsort (rela + ret, count - ret, sizeof (*rela), elf_link_sort_cmp2);
+  qsort (rela + ret, (size_t) count - ret, sizeof (*rela), elf_link_sort_cmp2);
 
   for (o = dynobj->sections; o != NULL; o = o->next)
     if ((o->flags & (SEC_HAS_CONTENTS|SEC_LINKER_CREATED))
